parrishstreet/requirements/base.txt
Django==1.3.1
Fabric==1.4.1
Markdown==2.1.1
PIL==1.1.7
Pinax==0.9a2
South==0.7.4
django-admin-tools==0.4.1
django-announcements==0.1.2
django-appconf==0.5
django-compressor==1.0.1
django-debug-toolbar==0.8.5
django-email-confirmation==0.2
django-mailer==0.2a1
django-notification==0.2
django-openid==0.3a1
django-pagination==1.0.7
django-staticfiles==1.1.2
django-tagging==0.3.1
django-timezones==0.2
geopy==0.94.2
idios==0.2a1
metron==0.1
pinax-theme-bootstrap==0.1.2
pinax-theme-foundation==0.1.4
psycopg2==2.4.5
pycrypto==2.5
python-openid==2.2.5
pytz==2011n
ssh==1.7.13
wsgiref==0.1.2
